move pkg/node/defaults.go to pkg/defaults/node.go
This allows several packages to not have to use the `pkg/node` dependency, which
pulls in `pkg/bpf`, which uses the runtime/cgo dependency. This is useful
because it removes the need for glibc to be present for some packages, like
`pkg/endpoint/connector`.

Fixes: #5055

Signed-off by: Ian Vernon <ian@cilium.io>
